Changes the Nukkit's HeightMap to ByteArray and fills it with 255. Fixes #39

+++ src/main/kotlin/br/com/gamemods/j2nwc/internal/NukkitChunk.kt
@@ -57,7 +57,7 @@ internal data class NukkitChunk(
     var position: ChunkPos,
     var biomes: ByteArray,
     var extraData: ByteArray,
-    var heightMap: IntArray
+    var heightMap: ByteArray
 ) {
     fun toNbt(): NbtFile {
         val level = NbtCompound()
